+++ backend/tessend/tessend.go
@@ -241,6 +241,7 @@ func (t *Tessend) startMessageHandler() {
 					}
 					metric.Tags = append(metric.Tags, &corev2.MetricTag{Name: "hostname", Value: hostname})
 					metric.Timestamp = now
+					appendInternalTag(&metric)
 					logMetric(&metric)
 					data.Metrics.Points = append(data.Metrics.Points, &metric)
 				}
@@ -401,6 +402,7 @@ func (t *Tessend) sendPromMetrics() {
 			},
 		},
 	}
+	appendInternalTag(mp)
 	logMetric(mp)
 	data.Metrics.Points = append(data.Metrics.Points, mp)
 
@@ -545,6 +547,7 @@ func (t *Tessend) getPerResourceMetrics(now int64, data *Data) {
 		Value:     backendCount,
 		Timestamp: now,
 	}
+	appendInternalTag(mp)
 	logMetric(mp)
 	data.Metrics.Points = append(data.Metrics.Points, mp)
 
@@ -563,6 +566,7 @@ func (t *Tessend) getPerResourceMetrics(now int64, data *Data) {
 			Value:     float64(count),
 			Timestamp: now,
 		}
+		appendInternalTag(mp)
 		logMetric(mp)
 		data.Metrics.Points = append(data.Metrics.Points, mp)
 	}
@@ -581,6 +585,7 @@ func (t *Tessend) getTessenConfigMetrics(now int64, tessen *corev2.TessenConfig,
 			},
 		},
 	}
+	appendInternalTag(mp)
 	logMetric(mp)
 	data.Metrics.Points = append(data.Metrics.Points, mp)
 }
@@ -610,4 +615,14 @@ func logMetric(m *corev2.MetricPoint) {
 		"metric_name":  m.Name,
 		"metric_value": m.Value,
 	}).Debug("collected a metric for tessen")
+}
+
+// appendInternalTag tags the metric with an internal environment variable value
+func appendInternalTag(m *corev2.MetricPoint) {
+	if internalEnv := os.Getenv("SENSU_INTERNAL_ENVIRONMENT"); internalEnv != "" {
+		m.Tags = append(m.Tags, &corev2.MetricTag{
+			Name:  "sensu_internal_environment",
+			Value: internalEnv,
+		})
+	}
 }